+++ changelog.markdown
@@ -1,4 +1,11 @@
 # Changelog
+## 1.8.0
+* When the subject's constructor throws an exception, it is now bubbled up and shown in the failure message.
+* Fixed an exception in the Rhino Mocks adapter when faking a delegate (thanks to [Alexis Atkinson](https://github.com/alexisatkinson)).
+* Updated to FakeItEasy 1.14.0
+* Updated to Machine.Specifications 0.5.16
+* Updated to Moq 4.1.1309.1617
+
 ## 1.7.0
 * Added support for setting up values for out and ref parameters to FakeItEasy and NSubsitute adapters.
 * Updated to Machine.Specifications 0.5.14